+++ ci/gpu/build.sh
@@ -26,6 +26,7 @@ export CUDA_REL=${CUDA_VERSION%.*}
 export HOME=$WORKSPACE
 
 # Parse git describe
+cd $WORKSPACE
 export GIT_DESCRIBE_TAG=`git describe --tags`
 export MINOR_VERSION=`echo $GIT_DESCRIBE_TAG | grep -o -E '([0-9]+\.[0-9]+)'`